Add section in README to stop or quit the process.
I had to go hunting for this and found it [here](https://github.com/bravobit/FFmpeg-Android/issues/4) and [here](https://github.com/bravobit/FFmpeg-Android/wiki/Send-'Q'-signal-to-FFmpeg-FFprobe). This is a fairly common use case, especially for people streaming from an RTSP or similar, so let's put it in the README and save future people time, energy and frustration.